Add files via upload
The file now concatenates all of the models' outputs as classifiers in two dataframes: train_df and test_df.  
The next step for the ensemble is to train the final model on train_df and y_train, and then evaluate the model with test_df and y_test.
All data has been prepared.