Resampled Data before running through models
This file tries to improve the performance of Linear, Decision Tree, and Support Vector models.  Logistic Regression was performing at 86% accuracy with regards to the target 'Bankrupt?'=1.  However, the other models were not sensitive to the feature, as 96% of the data had 'Bankrupt?'=0.  The models had become very biased towards 'Bankrupt?'=0.  Setting their classifications to all 0's yielded them an apparent accuracy score of 96%.  However, with regards to 'Bankrupt?'=1, those models ranged from 0% accuracy to 26% accuracy.  This file resamples the data twice before running it.  The resampling results in a significant improvement in the performance of Decision Tree, which goes from 0% to 58.82%, and Support Vector Machines I and II, which both improve from 0% to 62.75%.  The resampling reduced the performance of Logistic Regression from 86% to 58.82%, and Linear regression fell from 26% to 2.5%.  Additional changes would be needed in order to preserve the performance of Linear and Logistic Regression, while improving the performance of Decision Tree and Support Vector Machine.  The results from these models were then fed to an artificial neural network as input nodes.  The first ANN takes the straight data and results in 60.78% accuracy.  The second resamples the data from the classifiers and arrives at the same accuracy: 60.78%.  The third resamples again and improve 4% points to arrive at 64.71% accuracy.  None of these techniques reaches the performance of the Logistic Regression model, with no resampling of the data, which was able to achieve 86% accuracy.